fix(Component Service Lookup Service): Remove circular dependency
The circular dependency occurred because
AnnotationService > ProjectService > ComponentServiceLookupService >
EmbeddedService > AnnotationService

#637
