+++ src/assets/wise5/components/embedded/embeddedService.ts
@@ -7,7 +7,6 @@ import { StudentAssetService } from '../../services/studentAssetService';
 import { Injectable } from '@angular/core';
 import { UtilService } from '../../services/utilService';
 import { ConfigService } from '../../services/configService';
-import { AnnotationService } from '../../services/annotationService';
 
 @Injectable()
 export class EmbeddedService extends ComponentService {
@@ -16,7 +15,6 @@ export class EmbeddedService extends ComponentService {
   iframePrefix: string = 'embedded-application-iframe-';
 
   constructor(
-    protected AnnotationService: AnnotationService,
     protected ConfigService: ConfigService,
     protected StudentAssetService: StudentAssetService,
     protected UtilService: UtilService
@@ -142,29 +140,6 @@ export class EmbeddedService extends ComponentService {
     const message = {
       messageType: 'parameters',
       parameters: parameters
-    };
-    this.sendMessageToApplication(iframeId, message);
-  }
-
-  handleGetLatestAnnotationsMessage(iframeId: string, nodeId: string, componentId: string): void {
-    const workgroupId = this.ConfigService.getWorkgroupId();
-    const type = 'any';
-    const latestScoreAnnotation = this.AnnotationService.getLatestScoreAnnotation(
-      nodeId,
-      componentId,
-      workgroupId,
-      type
-    );
-    const latestCommentAnnotation = this.AnnotationService.getLatestCommentAnnotation(
-      nodeId,
-      componentId,
-      workgroupId,
-      type
-    );
-    const message = {
-      messageType: 'latestAnnotations',
-      latestScoreAnnotation: latestScoreAnnotation,
-      latestCommentAnnotation: latestCommentAnnotation
     };
     this.sendMessageToApplication(iframeId, message);
   }
